--- a/pkg/types/upstream.go
+++ b/pkg/types/upstream.go
@@ -113,9 +113,6 @@ type HostPredicate func(Host) bool
 type HostSet interface {
 	// Hosts returns all hosts that make up the set at the current time.
 	Hosts() []Host
-
-	// HealthyHosts returns all healthy hosts
-	HealthyHosts() []Host
 }
 
 // Host is an upstream host