--- a/pkg/upstream/cluster/host_set.go
+++ b/pkg/upstream/cluster/host_set.go
@@ -29,10 +29,6 @@ type hostSet struct {
 	once     sync.Once
 	mux      sync.RWMutex
 	allHosts []types.Host
-	// if refresh is true, the healthyHosts needs to refresh
-	refresh       bool
-	refreshNotify []func(host types.Host)
-	healthyHosts  []types.Host
 }
 
 // Hosts do not needs lock, becasue it "immutable"
@@ -40,64 +36,18 @@ func (hs *hostSet) Hosts() []types.Host {
 	return hs.allHosts
 }
 
-func (hs *hostSet) HealthyHosts() []types.Host {
-	hs.mux.RLock()
-	defer hs.mux.RUnlock()
-	return hs.healthyHosts
-}
-
-// refresh notify do not needs lock, the createSubset will not be called in concurrency
-// once the refreshNotify slice have been created, it will not be changed
-func (hs *hostSet) addRefreshNotify(f func(host types.Host)) {
-	hs.refreshNotify = append(hs.refreshNotify, f)
-}
-
-func (hs *hostSet) getRefreshNotify() []func(host types.Host) {
-	return hs.refreshNotify
-}
-
-func (hs *hostSet) resetHealthyHosts() {
-	healthyHosts := make([]types.Host, 0, len(hs.allHosts))
-	for _, h := range hs.allHosts {
-		if h.Health() {
-			healthyHosts = append(healthyHosts, h)
-		}
-	}
-	hs.mux.Lock()
-	defer hs.mux.Unlock()
-	hs.healthyHosts = healthyHosts
-
-}
-
-// refreshHealthHost resetHealthyHosts, and send a notify
-func (hs *hostSet) refreshHealthHost(host types.Host) {
-	hs.resetHealthyHosts()
-	// send refresh notify
-	refreshs := hs.getRefreshNotify()
-	for _, refresh := range refreshs {
-		refresh(host)
-	}
-}
-
 func (hs *hostSet) createSubset(predicate types.HostPredicate) types.HostSet {
 	allHosts := hs.Hosts()
 	var subHosts []types.Host
-	var healthyHosts []types.Host
 	for _, h := range allHosts {
 		if predicate(h) {
 			subHosts = append(subHosts, h)
-			if h.Health() {
-				healthyHosts = append(healthyHosts, h)
-			}
 		}
 	}
 	sub := &subHostSet{
-		predicate:    predicate,
-		allHosts:     subHosts,
-		healthyHosts: healthyHosts,
+		predicate: predicate,
+		allHosts:  subHosts,
 	}
-	// register refresh notify
-	hs.addRefreshNotify(sub.refresh)
 	return sub
 }
 
@@ -117,7 +67,6 @@ func (hs *hostSet) setFinalHost(hosts []types.Host) {
 			allHosts = append(allHosts, h)
 		}
 		hs.allHosts = allHosts
-		hs.resetHealthyHosts()
 		if log.DefaultLogger.GetLogLevel() >= log.INFO {
 			log.DefaultLogger.Infof("[upstream] [host set] update host, final host total: %d", len(hs.allHosts))
 		}
@@ -128,39 +77,12 @@ func (hs *hostSet) setFinalHost(hosts []types.Host) {
 // the member and healthy states is sync from hostSet that created it
 // no callbacks can be added in subset hostset
 type subHostSet struct {
-	mux          sync.RWMutex
-	allHosts     []types.Host
-	healthyHosts []types.Host
-	predicate    types.HostPredicate
+	mux       sync.RWMutex
+	allHosts  []types.Host
+	predicate types.HostPredicate
 }
 
 // Hosts do not need lock, because it is "immutable"
 func (sub *subHostSet) Hosts() []types.Host {
 	return sub.allHosts
-}
-
-func (sub *subHostSet) HealthyHosts() []types.Host {
-	sub.mux.RLock()
-	defer sub.mux.RUnlock()
-
-	return sub.healthyHosts
-}
-
-func (sub *subHostSet) resetHealthyHosts() {
-	healthyHosts := make([]types.Host, 0, len(sub.allHosts))
-	for _, h := range sub.allHosts {
-		if h.Health() {
-			healthyHosts = append(healthyHosts, h)
-		}
-	}
-	sub.mux.Lock()
-	defer sub.mux.Unlock()
-	sub.healthyHosts = healthyHosts
-}
-
-func (sub *subHostSet) refresh(host types.Host) {
-	if !sub.predicate(host) {
-		return
-	}
-	sub.resetHealthyHosts()
 }