--- a/pkg/upstream/cluster/host_set_test.go
+++ b/pkg/upstream/cluster/host_set_test.go
@@ -18,7 +18,6 @@
 package cluster
 
 import (
-	"fmt"
 	"testing"
 
 	"mosn.io/api"
@@ -52,68 +51,5 @@ func TestHostSetDistinct(t *testing.T) {
 	hs.setFinalHost(hosts)
 	if len(hs.Hosts()) != 1 {
 		t.Fatal("hostset distinct failed")
-	}
-}
-
-// Test Refresh healthy hosts
-func TestHostSetRefresh(t *testing.T) {
-	hs := &hostSet{}
-	configs := []simpleMockHostConfig{}
-	for i := 10000; i < 10010; i++ {
-		cfg := simpleMockHostConfig{
-			addr:      fmt.Sprintf("127.0.0.1:%d", i),
-			metaValue: "v1",
-		}
-		configs = append(configs, cfg)
-	}
-	for i := 11000; i < 11010; i++ {
-		cfg := simpleMockHostConfig{
-			addr:      fmt.Sprintf("127.0.0.1:%d", i),
-			metaValue: "v2",
-		}
-		configs = append(configs, cfg)
-	}
-	var hosts []types.Host
-	for _, cfg := range configs {
-		h := newSimpleMockHost(cfg.addr, cfg.metaValue)
-		hosts = append(hosts, h)
-	}
-	hs.setFinalHost(hosts)
-	if !(len(hs.allHosts) == len(hs.healthyHosts) &&
-		len(hs.allHosts) == 20) {
-		t.Fatalf("update host set not expected, %v", hs)
-	}
-	// create subset
-	subV1 := hs.createSubset(func(host types.Host) bool {
-		if host.Metadata()["key"] == "v1" {
-			return true
-		}
-		return false
-	})
-	subV2 := hs.createSubset(func(host types.Host) bool {
-		if host.Metadata()["key"] == "v2" {
-			return true
-		}
-		return false
-	})
-	// verify subv1 and subv2
-	if !(len(subV1.Hosts()) == 10 &&
-		len(subV2.Hosts()) == 10) {
-		t.Fatalf("create sub host set failed")
-	}
-	for _, h := range subV1.Hosts() {
-		if h.Metadata()["key"] != "v1" {
-			t.Fatal("sub host set v1 got unepxected host")
-		}
-	}
-	// mock health check set
-	allHosts := hs.Hosts()
-	host := allHosts[5] // mock one host healthy is changed
-	host.SetHealthFlag(api.FAILED_ACTIVE_HC)
-	hs.refreshHealthHost(host)
-	if !(len(hs.HealthyHosts()) == 19 &&
-		len(subV1.HealthyHosts()) == 9 &&
-		len(subV2.HealthyHosts()) == 10) {
-		t.Fatal("health check state changed not expected")
 	}
 }