delete healthy hosts, lb use hosts and check wether the host is healthy.

benchmark (new)
```
BenchmarkHostConfig/Host.Config-4          	2000000000	         0.28 ns/op
BenchmarkUpdateClusterHosts/UpdateClusterHost-4         	    3000	    422221 ns/op
BenchmarkClusterManagerRemoveHosts/ClusterRemoveHosts-4 	    2000	   1044833 ns/op
BenchmarkClusterManagerAppendHosts/ClusterManagerAppendHosts-4         	    1000	   1371648 ns/op
BenchmarkRandomLB-4                                                    	10000000	       122 ns/op
BenchmarkRandomLBWithUnhealthyHost-4                                   	10000000	       146 ns/op
BenchmarkRoundRobinLB-4                                                	50000000	        31.3 ns/op
BenchmarkRoundRobinLBWithUnhealthyHost-4                               	30000000	        52.1 ns/op
BenchmarkSubsetLB/CtxZone-4                                            	10000000	       203 ns/op
BenchmarkSubsetLB/CtxZoneAndVersion-4                                  	 5000000	       212 ns/op
BenchmarkSubsetLB/CtxNotMatched-4                                      	10000000	       195 ns/op
```

The odler
```
BenchmarkHostSetRefresh/RefreshSimple100-4         	  500000	      2960 ns/op
BenchmarkHostSetRefresh/RefreshSimple1000-4        	   50000	     29851 ns/op
BenchmarkHostSetRefresh/RefreshMeta1000-4          	   30000	     59985 ns/op
BenchmarkHostConfig/Host.Config-4                  	2000000000	         0.61 ns/op
BenchmarkUpdateClusterHosts/UpdateClusterHost-4    	    3000	    489080 ns/op
BenchmarkClusterManagerRemoveHosts/ClusterRemoveHosts-4         	    2000	   1029225 ns/op
BenchmarkClusterManagerAppendHosts/ClusterManagerAppendHosts-4  	    2000	    993200 ns/op
BenchmarkRandomLB-4                                             	10000000	       163 ns/op
BenchmarkRoundRobinLB-4                                         	20000000	        64.7 ns/op
BenchmarkSubsetLB/CtxZone-4                                     	10000000	       201 ns/op
BenchmarkSubsetLB/CtxZoneAndVersion-4                           	10000000	       225 ns/op
BenchmarkSubsetLB/CtxNotMatched-4                               	10000000	       209 ns/op
```